add delete update create card

+++ src/utils/deleteFile.ts
@@ -1,8 +1,10 @@
 import { deleteObject, ref } from "firebase/storage";
 import { storage } from "../firebase/config";
+import { checkFileExistence } from "./checkFileExistence";
 
 export const deleteFile = async (filename: string) => {
-	if (filename) {
+	const isImageExist = await checkFileExistence(filename);
+	if (filename && isImageExist) {
 		const fileRef = ref(storage, filename);
 		await deleteObject(fileRef);
 	}
